Reworked timer system to be more flexible. New RobotEvent methods to start in different periods. New Timer class to handle main timer. Separate functions for each match period timer. Timestamps changed to Long type for future-proofing. Some code in preparation of future update; time difference

+++ PracticeApp/src/main/java/com/knotbot/practiceapp/DataStorage.java
@@ -60,7 +60,7 @@ public class DataStorage {
 		if (runData.timestamp == null) {
 			saveRun(runData, "null");
 		} else {
-			saveRun(runData, Integer.toString(runData.timestamp));
+			saveRun(runData, Long.toString(runData.timestamp));
 		}
 	}
